refactor: Move App to components
