+++ src/components/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Practice2 from './components/practice2/Practice2';
+import Practice2 from './practice2/Practice2';
 
 function App() {
   return (
